Move anidb34 & tvdb4 functions into their own modules
